Safari-609.1.11.1/LayoutTests/js/function-declaration-expected.txt

This test ensures we correctly parse (or not) various function declarations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS eval('function f(){return true;}') is undefined.
PASS eval('function f(){return true;};f')() is true
PASS eval('function(){return false;}')() threw exception SyntaxError: Function statements must have a name..
PASS successfullyParsed is true

TEST COMPLETE
